requirements.txt
# pytorch 1.13.1 cuda 11.6
matplotlib==3.7.1
numpy==1.24.3
opencv_python==4.7.0.72
pandas==1.5.3
Pillow==9.5.0
seaborn==0.12.2
skimage==0.0
tqdm==4.65.0
ultralytics==8.0.53
